refactor: 메뉴 삭제 버튼 위치 이동
