+++ src/page/AddMenu/AddMenu.module.scss
@@ -28,6 +28,7 @@
     width: 100%;
     padding: 0 16px;
     margin-top: 16px;
+    margin-bottom: 16px;
     display: flex;
     align-items: center;
     justify-content: space-between;
