+++ src/page/ModifyMenu/index.tsx
@@ -93,11 +93,6 @@ export default function ModifyMenu() {
     openGoMyShopModal();
   };
 
-  const handleMobileDeleteMenu = () => {
-    deleteMenuMutation(Number(menuId));
-    goMyShop();
-  };
-
   return (
     <div>
       {isMobile ? (
@@ -161,15 +156,6 @@ export default function ModifyMenu() {
               </>
             )}
 
-          </div>
-          <div className={styles['mobile-delete-menu-container']}>
-            <button
-              className={styles['delete-menu-button']}
-              type="button"
-              onClick={handleMobileDeleteMenu}
-            >
-              메뉴 삭제
-            </button>
           </div>
         </div>
       ) : (
